+++ FairyPlugin/src/main/java/com/limpoxe/fairy/core/PluginIntentResolver.java
@@ -192,7 +192,7 @@ public class PluginIntentResolver {
 			PluginActivityInfo pluginActivityInfo = pluginDescriptor.getActivityInfos().get(className);
 
 			String stubActivityName = PluginManagerProviderClient.bindStubActivity(className,
-					Integer.parseInt(pluginActivityInfo.getLaunchMode()),
+                    (int)Long.parseLong(pluginActivityInfo.getLaunchMode()),
 					pluginDescriptor.getPackageName(),
 					pluginActivityInfo.getTheme(),
                     pluginActivityInfo.getScreenOrientation());
@@ -200,7 +200,7 @@ public class PluginIntentResolver {
             if (stubActivityName == null) {
                 LogUtil.e("绑定StubAtivity失败",
                         className,
-                        Integer.parseInt(pluginActivityInfo.getLaunchMode()),
+                        (int)Long.parseLong(pluginActivityInfo.getLaunchMode()),
                         pluginDescriptor.getPackageName(),
                         pluginActivityInfo.getTheme(),
                         pluginActivityInfo.getScreenOrientation());
